Adding different types of elevaors
1. PassengerElevator
2. FreightElevator
